fix: Move previous NIP-20 code into NIP-01.
