+++ src/memorelay.ts
@@ -11,7 +11,6 @@ import { PluginFn } from './core/types/plugin-types';
 import { basicProtocol } from './nips/nip-0001-basic-protocol/plugins';
 import { eventDeletion } from './nips/nip-0009-event-deletion/plugins';
 import { relayInformationDocument } from './nips/nip-0011-relay-information-document/plugins/relay-information-document';
-import { commandResults } from './nips/nip-0020-command-results/plugins';
 
 /**
  * Memorelay main class. Extends MemorelayHub and attaches default behavior.
@@ -43,9 +42,6 @@ export class Memorelay extends MemorelayHub {
     super(
       ...plugins,
       ...[
-        // NIP-20 command results.
-        commandResults, // NIP-20 Command Results.
-
         // NIP-11 relay information document requests.
         relayInformationDocument,
