+++ src/nips/nip-0001-basic-protocol/events/outgoing-ok-message-event.ts
