+++ src/nips/nip-0001-basic-protocol/lib/check-ok-message.test.ts
