+++ src/nips/nip-0001-basic-protocol/plugins/send-ok-after-database-add.test.ts
@@ -2,29 +2,28 @@
  * @license SPDX-License-Identifier: Apache-2.0
  */
 /**
- * @fileoverview Tests for sendOKAfterDuplicate().
+ * @fileoverview Tests for sendOKAfterDatabaseAdd().
  */
 
 import { setupTestHubAndClient } from '../../../test/setup-test-hub-and-client';
 import { createSignedTestEvent } from '../../../test/signed-test-event';
-import { DuplicateEventMessageEvent } from '../../nip-0001-basic-protocol/events/duplicate-event-message-event';
+import { DidAddEventToDatabaseEvent } from '../events/did-add-event-to-database-event';
 import { OutgoingOKMessageEvent } from '../events/outgoing-ok-message-event';
-import { sendOKAfterDuplicate } from './send-ok-after-duplicate';
+import { sendOKAfterDatabaseAdd } from './send-ok-after-database-add';
 
-describe('sendOKAfterDuplicate()', () => {
-  describe('#DuplicateEventMessageEvent', () => {
+describe('sendOKAfterDatabaseAdd()', () => {
+  describe('#DidAddEventToDatabaseEvent', () => {
     it('should send an outgoing OK message event', async () => {
-      const { memorelayClient } = setupTestHubAndClient();
-      sendOKAfterDuplicate(memorelayClient);
+      const { memorelayClient } = setupTestHubAndClient(sendOKAfterDatabaseAdd);
 
       const mockHandlerFn = jest.fn<unknown, [OutgoingOKMessageEvent]>();
       memorelayClient.onEvent(OutgoingOKMessageEvent, mockHandlerFn);
 
       const testEvent = createSignedTestEvent({ content: 'TEST' });
-      const duplicateEventMessageEvent = new DuplicateEventMessageEvent({
+      const didAddEventToDatabaseEvent = new DidAddEventToDatabaseEvent({
         event: testEvent,
       });
-      memorelayClient.emitEvent(duplicateEventMessageEvent);
+      memorelayClient.emitEvent(didAddEventToDatabaseEvent);
 
       expect(mockHandlerFn).not.toHaveBeenCalled();
 
@@ -35,29 +34,22 @@ describe('sendOKAfterDuplicate()', () => {
       const outgoingOKMessageEvent = mockHandlerFn.mock.calls[0][0];
       expect(outgoingOKMessageEvent).toBeInstanceOf(OutgoingOKMessageEvent);
       expect(outgoingOKMessageEvent.parentEvent).toBe(
-        duplicateEventMessageEvent
+        didAddEventToDatabaseEvent
       );
-      expect(outgoingOKMessageEvent.details.okMessage).toEqual([
-        'OK',
-        testEvent.id,
-        true,
-        'duplicate:',
-      ]);
     });
 
     it('should not send when defaultPrevented', async () => {
-      const { memorelayClient } = setupTestHubAndClient();
-      sendOKAfterDuplicate(memorelayClient);
+      const { memorelayClient } = setupTestHubAndClient(sendOKAfterDatabaseAdd);
 
       const mockHandlerFn = jest.fn<unknown, [OutgoingOKMessageEvent]>();
       memorelayClient.onEvent(OutgoingOKMessageEvent, mockHandlerFn);
 
       const testEvent = createSignedTestEvent({ content: 'TEST' });
-      const duplicateEventMessageEvent = new DuplicateEventMessageEvent({
+      const didAddEventToDatabaseEvent = new DidAddEventToDatabaseEvent({
         event: testEvent,
       });
-      duplicateEventMessageEvent.preventDefault();
-      memorelayClient.emitEvent(duplicateEventMessageEvent);
+      didAddEventToDatabaseEvent.preventDefault();
+      memorelayClient.emitEvent(didAddEventToDatabaseEvent);
 
       await Promise.resolve();
