+++ src/nips/nip-0001-basic-protocol/plugins/send-ok-after-duplicate.test.ts
@@ -2,29 +2,28 @@
  * @license SPDX-License-Identifier: Apache-2.0
  */
 /**
- * @fileoverview Tests for sendOKAfterDatabaseAdd().
+ * @fileoverview Tests for sendOKAfterDuplicate().
  */
 
 import { setupTestHubAndClient } from '../../../test/setup-test-hub-and-client';
 import { createSignedTestEvent } from '../../../test/signed-test-event';
-import { DidAddEventToDatabaseEvent } from '../../nip-0001-basic-protocol/events/did-add-event-to-database-event';
+import { DuplicateEventMessageEvent } from '../events/duplicate-event-message-event';
 import { OutgoingOKMessageEvent } from '../events/outgoing-ok-message-event';
-import { sendOKAfterDatabaseAdd } from './send-ok-after-database-add';
+import { sendOKAfterDuplicate } from './send-ok-after-duplicate';
 
-describe('sendOKAfterDatabaseAdd()', () => {
-  describe('#DidAddEventToDatabaseEvent', () => {
+describe('sendOKAfterDuplicate()', () => {
+  describe('#DuplicateEventMessageEvent', () => {
     it('should send an outgoing OK message event', async () => {
-      const { memorelayClient } = setupTestHubAndClient();
-      sendOKAfterDatabaseAdd(memorelayClient);
+      const { memorelayClient } = setupTestHubAndClient(sendOKAfterDuplicate);
 
       const mockHandlerFn = jest.fn<unknown, [OutgoingOKMessageEvent]>();
       memorelayClient.onEvent(OutgoingOKMessageEvent, mockHandlerFn);
 
       const testEvent = createSignedTestEvent({ content: 'TEST' });
-      const didAddEventToDatabaseEvent = new DidAddEventToDatabaseEvent({
+      const duplicateEventMessageEvent = new DuplicateEventMessageEvent({
         event: testEvent,
       });
-      memorelayClient.emitEvent(didAddEventToDatabaseEvent);
+      memorelayClient.emitEvent(duplicateEventMessageEvent);
 
       expect(mockHandlerFn).not.toHaveBeenCalled();
 
@@ -35,23 +34,28 @@ describe('sendOKAfterDatabaseAdd()', () => {
       const outgoingOKMessageEvent = mockHandlerFn.mock.calls[0][0];
       expect(outgoingOKMessageEvent).toBeInstanceOf(OutgoingOKMessageEvent);
       expect(outgoingOKMessageEvent.parentEvent).toBe(
-        didAddEventToDatabaseEvent
+        duplicateEventMessageEvent
       );
+      expect(outgoingOKMessageEvent.details.okMessage).toEqual([
+        'OK',
+        testEvent.id,
+        true,
+        'duplicate:',
+      ]);
     });
 
     it('should not send when defaultPrevented', async () => {
-      const { memorelayClient } = setupTestHubAndClient();
-      sendOKAfterDatabaseAdd(memorelayClient);
+      const { memorelayClient } = setupTestHubAndClient(sendOKAfterDuplicate);
 
       const mockHandlerFn = jest.fn<unknown, [OutgoingOKMessageEvent]>();
       memorelayClient.onEvent(OutgoingOKMessageEvent, mockHandlerFn);
 
       const testEvent = createSignedTestEvent({ content: 'TEST' });
-      const didAddEventToDatabaseEvent = new DidAddEventToDatabaseEvent({
+      const duplicateEventMessageEvent = new DuplicateEventMessageEvent({
         event: testEvent,
       });
-      didAddEventToDatabaseEvent.preventDefault();
-      memorelayClient.emitEvent(didAddEventToDatabaseEvent);
+      duplicateEventMessageEvent.preventDefault();
+      memorelayClient.emitEvent(duplicateEventMessageEvent);
 
       await Promise.resolve();
